// planning-idsdp-rms-frontend/src/app/main/modules/training-institute/models/nominated-institute.model.ts
export class NominatedInstituteModel {
    trainingName: string;
    instituteName: string;
    trainingDuration: number;
    programDate: Date;
    fiscalYearId: number;
    trainingInstituteName: string;
    trainingInstituteNameDetails: string;
    proposalStatus: string;
    isShortListed: boolean;
    editable: boolean;
    submitted: boolean;
    id: number;
    isCompletionReportSubmitted: boolean;
}